step04export/module/requirements.txt
numpy>=2.4.0
onnxruntime-gpu; sys_platform != "darwin"
onnxruntime; sys_platform == "darwin"
transformers
sentence-transformers
scikit-learn
Pillow
torch
lightgbm